requirements.txt
pillow==9.5.0
albumentations==0.4.3
opencv-python #==4.1.2.30
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
transformers==4.36.0
torchmetrics==0.7.0
kornia==0.6.4
#-e git+https://github.com/arogozhnikov/einops.git@main#egg=einops